fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,17 +12,18 @@
 int main(void)
 {
 	int i;
-	int fib[N];
+	unsigned long int fib[N];
+
 	fib[0] = 1;
 	fib[1] = 2;
 
 	for (i = 2; i < N; i++)
 	{
-		fib[i] = fib[i-1] + fib[i-2];
+		fib[i] = fib[i - 1] + fib[i - 2];
 	}
 	for (i = 0; i < N; i++)
 	{
-		printf("%d", fib[i]);
+		printf("%lu", fib[i]);
 		if (i < N - 1)
 		{
 			printf(", ");
